+++ packages/docs/src/contexts/Wallet.tsx
@@ -1,39 +1,42 @@
-import { WalletAdapterNetwork } from "@solana/wallet-adapter-base";
+import React, { FC, useMemo } from "react";
 import {
   ConnectionProvider,
   WalletProvider,
 } from "@solana/wallet-adapter-react";
 import {
-  getLedgerWallet,
-  getPhantomWallet,
-  getSlopeWallet,
-  getSolflareWallet,
-  getSolletExtensionWallet,
-  getSolletWallet,
+  LedgerWalletAdapter,
+  PhantomWalletAdapter,
+  SlopeWalletAdapter,
+  SolflareWalletAdapter,
+  SolletExtensionWalletAdapter,
+  SolletWalletAdapter,
+  TorusWalletAdapter,
 } from "@solana/wallet-adapter-wallets";
-import React, { FC, useMemo } from "react";
 import { useEndpoint } from "./Endpoint";
 
 // Default styles that can be overridden by your app
 require("@solana/wallet-adapter-react-ui/styles.css");
 
-export const Wallet = ({ children }: { children?: React.ReactNode }) => {
+export const Wallet: FC = ({ children }) => {
   // You can also provide a custom RPC endpoint
   const { endpoint } = useEndpoint();
 
-  // @solana/wallet-adapter-wallets includes all the adapters but supports tree shaking --
-  // Only the wallets you configure here will be compiled into your application
+  // @solana/wallet-adapter-wallets includes all the adapters but supports tree shaking and lazy loading --
+  // Only the wallets you configure here will be compiled into your application, and only the dependencies
+  // of wallets that your users connect to will be loaded
   const wallets = useMemo(
     () => [
-      getPhantomWallet(),
-      getSlopeWallet(),
-      getSolflareWallet(),
-      getLedgerWallet(),
-      getSolletWallet({}),
-      getSolletExtensionWallet({}),
+      new PhantomWalletAdapter(),
+      new SlopeWalletAdapter(),
+      new SolflareWalletAdapter(),
+      new TorusWalletAdapter(),
+      new LedgerWalletAdapter(),
+      new SolletWalletAdapter({}),
+      new SolletExtensionWalletAdapter({}),
     ],
     []
   );
+
 
   return (
     <ConnectionProvider endpoint={endpoint}>