+++ packages/react/src/contexts/accountContext.tsx
@@ -11,25 +11,30 @@ export interface IAccountProviderProps {
 }
 
 export const AccountContext = createContext<AccountFetchCache>(
-  {} as AccountFetchCache
+  // @ts-ignore
+  null as AccountFetchCache
 );
 
 export const AccountProvider: FC<IAccountProviderProps> = ({
   children,
   commitment = DEFAULT_COMMITMENT,
-  extendConnection = true
+  extendConnection = true,
 }) => {
   const { connection } = useConnection();
   const cache = useMemo(() => {
-    return new AccountFetchCache({
+    return connection && new AccountFetchCache({
       connection,
       delay: 500,
       commitment,
-      extendConnection
+      extendConnection,
     });
   }, [connection]);
 
-  return (
-    <AccountContext.Provider value={cache}>{children}</AccountContext.Provider>
-  );
+  if (cache) {
+    return (
+      <AccountContext.Provider value={cache}>{children}</AccountContext.Provider>
+    );
+  }
+
+  return null;
 };