+++ packages/react/src/index.ts
@@ -2,3 +2,4 @@ import "./bufferFill";
 export * from "./contexts";
 export * from "./hooks";
 export * from "./components";
+export * from "./utils";